--- src/services/AuthService.ts
@@ -1,23 +1,0 @@
-import { axiosInstance } from '../apis/axiosInstance/axiosInstance';
-
-export interface SignupRequest {
-  email: string;
-  password: string;
-  nickname: string;
-  phone: string;
-}
-
-export const signup = async (data: SignupRequest): Promise<any> => {
-  const response = await axiosInstance.post('/signup', data);
-  return response.data;
-};
-
-export const checkEmailDuplication = async (email: string): Promise<any> => {
-  const response = await axiosInstance.post('/emailcheck', { email });
-  return response.data;
-};
-
-export const checkNicknameDuplication = async (nickname: string): Promise<any> => {
-  const response = await axiosInstance.post('/nickcheck', { nickname });
-  return response.data;
-};
